devtools: Separate actors into their own files
Switch the actors to the register pattern.
