+++ components/devtools/actors/inspector/accessibility.rs
@@ -11,6 +11,8 @@ use serde_json::{Map, Value};
 
 use crate::StreamId;
 use crate::actor::{Actor, ActorError, ActorRegistry};
+use crate::actors::inspector::accessible_walker::AccessibleWalkerActor;
+use crate::actors::inspector::simulator::SimulatorActor;
 use crate::protocol::ClientRequest;
 
 #[derive(Serialize)]
@@ -90,14 +92,11 @@ impl Actor for AccessibilityActor {
                 request.reply_final(&msg)?
             },
             "getSimulator" => {
-                // TODO: Create actual simulator
-                let actor = registry.new_name::<SimulatorActor>();
-                registry.register(SimulatorActor {
-                    name: actor.clone(),
-                });
                 let msg = GetSimulatorReply {
                     from: self.name(),
-                    simulator: ActorMsg { actor },
+                    simulator: ActorMsg {
+                        actor: SimulatorActor::register(registry),
+                    },
                 };
                 request.reply_final(&msg)?
             },
@@ -111,11 +110,11 @@ impl Actor for AccessibilityActor {
                 request.reply_final(&msg)?
             },
             "getWalker" => {
-                // TODO: Create actual accessible walker
-                let actor = AccessibleWalkerActor::register(registry);
                 let msg = GetWalkerReply {
                     from: self.name(),
-                    walker: ActorMsg { actor },
+                    walker: ActorMsg {
+                        actor: AccessibleWalkerActor::register(registry),
+                    },
                 };
                 request.reply_final(&msg)?
             },
@@ -131,36 +130,5 @@ impl AccessibilityActor {
         let actor = Self { name: name.clone() };
         registry.register::<Self>(actor);
         name
-    }
-}
-
-#[derive(MallocSizeOf)]
-pub(crate) struct SimulatorActor {
-    name: String,
-}
-
-impl Actor for SimulatorActor {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-}
-
-#[derive(MallocSizeOf)]
-pub(crate) struct AccessibleWalkerActor {
-    name: String,
-}
-
-impl AccessibleWalkerActor {
-    pub fn register(registry: &ActorRegistry) -> String {
-        let name = registry.new_name::<Self>();
-        let actor = Self { name: name.clone() };
-        registry.register::<Self>(actor);
-        name
-    }
-}
-
-impl Actor for AccessibleWalkerActor {
-    fn name(&self) -> String {
-        self.name.clone()
     }
 }
