+++ components/devtools/actors/inspector/accessible_walker.rs
@@ -1,0 +1,27 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+use malloc_size_of_derive::MallocSizeOf;
+
+use crate::actor::{Actor, ActorRegistry};
+
+#[derive(MallocSizeOf)]
+pub(crate) struct AccessibleWalkerActor {
+    name: String,
+}
+
+impl AccessibleWalkerActor {
+    pub fn register(registry: &ActorRegistry) -> String {
+        let name = registry.new_name::<Self>();
+        let actor = Self { name: name.clone() };
+        registry.register::<Self>(actor);
+        name
+    }
+}
+
+impl Actor for AccessibleWalkerActor {
+    fn name(&self) -> String {
+        self.name.clone()
+    }
+}
